# openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AliExpress Affiliate API Proxy",
    "description": "Production-ready Python service for AliExpress Affiliate API with comprehensive product search, category management, and affiliate link generation capabilities. Perfect for custom GPT integration.",
    "version": "1.0.0",
    "contact": {
      "name": "AliExpress API Support",
      "url": "https://github.com/your-username/aliexpress-python-refactor"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-deployment-url.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/categories": {
      "get": {
        "summary": "Get Parent Categories",
        "description": "Retrieve all parent categories from AliExpress",
        "operationId": "getParentCategories",
        "responses": {
          "200": {
            "description": "List of parent categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "category_id": "509",
                      "category_name": "Electronics"
                    },
                    {
                      "category_id": "1420",
                      "category_name": "Fashion"
                    }
                  ],
                  "metadata": {
                    "total_count": 2,
                    "processing_time_ms": 150
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{parent_id}/children": {
      "get": {
        "summary": "Get Child Categories",
        "description": "Retrieve child categories for a specific parent category",
        "operationId": "getChildCategories",
        "parameters": [
          {
            "name": "parent_id",
            "in": "path",
            "required": true,
            "description": "The parent category ID",
            "schema": {
              "type": "string",
              "example": "509"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of child categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "category_id": "50906",
                      "category_name": "Smartphones",
                      "parent_id": "509"
                    },
                    {
                      "category_id": "50907",
                      "category_name": "Laptops",
                      "parent_id": "509"
                    }
                  ],
                  "metadata": {
                    "parent_id": "509",
                    "total_count": 2,
                    "processing_time_ms": 120
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parent_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/details": {
      "post": {
        "summary": "Get Product Details",
        "description": "Get detailed information for specific products",
        "operationId": "getProductDetails",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDetailsRequest"
              },
              "example": {
                "product_ids": ["1005001234567890", "1005001234567891"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detailed product information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/affiliate/links": {
      "post": {
        "summary": "Generate Affiliate Links",
        "description": "Generate affiliate tracking links for product URLs",
        "operationId": "generateAffiliateLinks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffiliateLinkRequest"
              },
              "example": {
                "urls": [
                  "https://www.aliexpress.com/item/1005001234567890.html",
                  "https://www.aliexpress.com/item/1005001234567891.html"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated affiliate links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateLinkResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "original_url": "https://www.aliexpress.com/item/1005001234567890.html",
                      "affiliate_url": "https://s.click.aliexpress.com/e/_tracking_link_1",
                      "tracking_id": "gpt_chat",
                      "commission_rate": "30%"
                    }
                  ],
                  "metadata": {
                    "total_links": 1,
                    "processing_time_ms": 200
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/affiliate/link": {
      "get": {
        "summary": "Generate Single Affiliate Link",
        "description": "Generate an affiliate link for a single product URL",
        "operationId": "generateSingleAffiliateLink",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "Product URL to convert to affiliate link",
            "schema": {
              "type": "string",
              "format": "uri",
              "example": "https://www.aliexpress.com/item/1005001234567890.html"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated affiliate link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleAffiliateLinkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "description": "Unique category identifier",
            "example": "509"
          },
          "category_name": {
            "type": "string",
            "description": "Category display name",
            "example": "Electronics"
          },
          "parent_id": {
            "type": "string",
            "description": "Parent category ID (for child categories)",
            "example": "509"
          }
        },
        "required": ["category_id", "category_name"]
      },
      "CategoriesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "total_count": {
                "type": "integer",
                "example": 2
              },
              "parent_id": {
                "type": "string",
                "example": "509"
              },
              "processing_time_ms": {
                "type": "number",
                "example": 150
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "Product": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "description": "Unique product identifier",
            "example": "1005001234567890"
          },
          "product_title": {
            "type": "string",
            "description": "Product title",
            "example": "Wireless Bluetooth Headphones"
          },
          "product_url": {
            "type": "string",
            "format": "uri",
            "description": "Product URL (may be affiliate link)",
            "example": "https://s.click.aliexpress.com/e/_affiliate_link"
          },
          "price": {
            "type": "string",
            "description": "Current sale price",
            "example": "29.99"
          },
          "currency": {
            "type": "string",
            "description": "Price currency",
            "example": "USD"
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "Main product image URL",
            "example": "https://ae01.alicdn.com/image.jpg"
          },
          "commission_rate": {
            "type": "string",
            "description": "Affiliate commission rate",
            "example": "30%"
          },
          "original_price": {
            "type": "string",
            "description": "Original price before discount",
            "example": "59.99"
          },
          "discount": {
            "type": "string",
            "description": "Discount percentage",
            "example": "50%"
          },
          "evaluate_rate": {
            "type": "string",
            "description": "Product rating percentage",
            "example": "98.5%"
          },
          "orders_count": {
            "type": "integer",
            "description": "Number of orders",
            "example": 1500
          }
        },
        "required": ["product_id", "product_title", "product_url", "price", "currency"]
      },
...
      "ProductDetailsRequest": {
        "type": "object",
        "properties": {
          "product_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "description": "List of product IDs (max 20)",
            "example": ["1005001234567890", "1005001234567891"]
          }
        },
        "required": ["product_ids"]
      },
      "ProductDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Product"
          },
          {
            "type": "object",
            "properties": {
              "gallery_images": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                },
                "description": "Additional product images"
              },
              "description": {
                "type": "string",
                "description": "Product description"
              },
              "specifications": {
                "type": "object",
                "description": "Product specifications"
              },
              "shipping_info": {
                "type": "object",
                "description": "Shipping information"
              },
              "seller_info": {
                "type": "object",
                "description": "Seller information"
              }
            }
          }
        ]
      },
      "ProductDetailsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDetail"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "total_products": {
                "type": "integer",
                "example": 2
              },
              "processing_time_ms": {
                "type": "number",
                "example": 300
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "AffiliateLinkRequest": {
        "type": "object",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "maxItems": 50,
            "description": "List of product URLs to convert (max 50)",
            "example": [
              "https://www.aliexpress.com/item/1005001234567890.html"
            ]
          }
        },
        "required": ["urls"]
      },
      "AffiliateLink": {
        "type": "object",
        "properties": {
          "original_url": {
            "type": "string",
            "format": "uri",
            "description": "Original product URL",
            "example": "https://www.aliexpress.com/item/1005001234567890.html"
          },
          "affiliate_url": {
            "type": "string",
            "format": "uri",
            "description": "Generated affiliate URL",
            "example": "https://s.click.aliexpress.com/e/_tracking_link_1"
          },
          "tracking_id": {
            "type": "string",
            "description": "Affiliate tracking ID",
            "example": "gpt_chat"
          },
          "commission_rate": {
            "type": "string",
            "description": "Commission rate for this product",
            "example": "30%"
          }
        },
        "required": ["original_url", "affiliate_url", "tracking_id"]
      },
      "AffiliateLinkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AffiliateLink"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "total_links": {
                "type": "integer",
                "example": 1
              },
              "processing_time_ms": {
                "type": "number",
                "example": 200
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "SingleAffiliateLinkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/AffiliateLink"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "processing_time_ms": {
                "type": "number",
                "example": 150
              }
            }
          }
        },
        "required": ["success", "data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid request parameters"
          },
          "error_code": {
            "type": "string",
            "description": "Error code",
            "example": "VALIDATION_ERROR"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "example": "2025-11-05T13:51:02.609Z"
              },
              "request_id": {
                "type": "string",
                "example": "req_123456789"
              }
            }
          }
        },
        "required": ["success", "error"]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for enhanced access"
      }
    }
  },
  "security": [
    {},
    {
      "ApiKeyAuth": []
    }
  ],
...
}